+++ client/package.json
@@ -1,7 +1,7 @@
 {
   "name": "client",
   "version": "0.1.0",
-  "proxy": "http://server:8080/",
+  "proxy": "http://localhost:8080/",
   "private": true,
   "dependencies": {
     "@emotion/react": "^11.10.5",
